Finally set up to run tests
