--- x/tokenomics/keeper/keeper_test.go
@@ -1,80 +1,0 @@
-package keeper_test
-
-import (
-	"context"
-	"testing"
-	"time"
-
-	"cosmossdk.io/math"
-	"github.com/cosmos/cosmos-sdk/types"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/stretchr/testify/suite"
-
-	"github.com/pokt-network/poktroll/cmd/poktrolld/cmd"
-	keepertest "github.com/pokt-network/poktroll/testutil/keeper"
-	"github.com/pokt-network/poktroll/testutil/sample"
-	testsession "github.com/pokt-network/poktroll/testutil/session"
-	apptypes "github.com/pokt-network/poktroll/x/application/types"
-	prooftypes "github.com/pokt-network/poktroll/x/proof/types"
-	sessiontypes "github.com/pokt-network/poktroll/x/session/types"
-	sharedtypes "github.com/pokt-network/poktroll/x/shared/types"
-)
-
-const (
-	testServiceId = "svc1"
-	testSessionId = "mock_session_id"
-)
-const minExecutionPeriod = 5 * time.Second
-
-func init() {
-	cmd.InitSDKConfig()
-}
-
-type TestSuite struct {
-	suite.Suite
-
-	sdkCtx  sdk.Context
-	ctx     context.Context
-	keepers keepertest.TokenomicsModuleKeepers
-	claim   prooftypes.Claim
-	proof   prooftypes.Proof
-}
-
-func (s *TestSuite) SetupTest() {
-	supplierAddr := sample.AccAddress()
-	appAddr := sample.AccAddress()
-
-	s.keepers, s.ctx = keepertest.NewTokenomicsModuleKeepers(s.T())
-	s.sdkCtx = sdk.UnwrapSDKContext(s.ctx)
-
-	// Prepare a claim that can be inserted
-	s.claim = prooftypes.Claim{
-		SupplierAddress: supplierAddr,
-		SessionHeader: &sessiontypes.SessionHeader{
-			ApplicationAddress:      appAddr,
-			Service:                 &sharedtypes.Service{Id: testServiceId},
-			SessionId:               "session_id",
-			SessionStartBlockHeight: 1,
-			SessionEndBlockHeight:   testsession.GetSessionEndHeightWithDefaultParams(1),
-		},
-		RootHash: smstRootWithSum(69),
-	}
-
-	// Prepare a claim that can be inserted
-	s.proof = prooftypes.Proof{
-		SupplierAddress: s.claim.SupplierAddress,
-		SessionHeader:   s.claim.SessionHeader,
-		// ClosestMerkleProof
-	}
-
-	appStake := types.NewCoin("upokt", math.NewInt(1000000))
-	app := apptypes.Application{
-		Address: appAddr,
-		Stake:   &appStake,
-	}
-	s.keepers.SetApplication(s.ctx, app)
-}
-
-func TestSettleExpiringClaimsSuite(t *testing.T) {
-	suite.Run(t, new(TestSuite))
-}
