+++ x/tokenomics/module/abci.go
@@ -31,6 +31,7 @@ func EndBlocker(ctx sdk.Context, k keeper.Keeper) error {
 		logger.Error(fmt.Sprintf("could not update relay mining difficulty due to error %v", err))
 		return err
 	}
+	logger.Info(fmt.Sprintf("successfully updated the relay mining difficulty for %d services", len(relaysPerServiceMap)))
 
 	return nil
 }
